Added a simple rater control to the lib
A simple rater control that could contain up to 10 stars. One could define the dark and bright color of the selected stars. It's simple but something that was fun to create. The size will be defined by the no of stars and the height of the preferred size of the control.