Added the possibility to create custom mappings for the SixteenSegment control
Moved the segment definition from the skin to the control so that one could get the current default mapping from ascii to segments. With this option you could create your own ascii to segment mapping and apply it to the 16 segment control. So one could create custom lcd characters. Also removed the glow again from the segments because it looked not really good when the segment becomes bigger in size.