Added a StepIndicator control
Right now one could choose between 3 and 10 steps and there is no behavior. The idea of Tom was to fire events if the mouse will be clicked on the filled circles to use it as a step indicator in a wizard, where one could step back by clicking on the completed steps.