The SplitFlap control now supports also custom text arrays
If you take a look at real solari boards you will find out that they contain flaps that take a whole word instead of only characters. This will increase the speed of the board a lot. With this revision of the SplitFlap control one could set an Array of Strings that contain the list of possible values. That means one could for example define a list of citynames and use them for one SplitFlap control as the a list of destinations etc.