Modified the handling of GaugeModelEvents and StyleModelEvents. Enabled the setAreas, setSections, setTickmarkSections and setMarkers methods again in the Gauge class.
In former versions the GaugeModelEvent and StyleModelEvent was only fired when the GaugeModel or StyleModel changed completely (which was not needed because they are hold in a property anyway which fires events by it's own). Now I've added a fireModelEvent call in each setter in both model classes. That means as soon as a property in a model will changes it's value the model should fire an event now.