Added a visual histogram feature to the Radial gauge. This feature will be available as soon as JavaFX 2.2 will be released (at the moment you will only see the last measured value).
The histogram will be created by drawing a line each time the value of the gauge changes. You could define the color of the line and the number of minutes you would like to be visible in the histogram. The value that is measured now will be invisible after the given time in minutes. That means over time the older lines will slowly disappear. With this feature you will get a visual impression of the deviation of the measured values in a given time period.