Added a redrawTolerance that could be between 0…1.
If you have lots of gauges on your screen and you apply always the same values to the gauges the program will go through all the code even if the pointer doesn't change it's position. To avoid this overhead the redrawTolerance will be used to calculate a tolerance range (multiply the redrawTolerance with the current range of the gauge). If the new value is within the oldValue +- this tolerance range the whole drawing procedure will be used otherwise the routine will simply return.